refactor: split notifications to a new cdc handler
To overcome the ordering constraint of CDC, we separate the notifications to a new topic to allow much faster handling with spikes

WT-2048 #done
