+++ .infra/application.properties
@@ -13,10 +13,13 @@ debezium.source.plugin.name=pgoutput
 debezium.source.heartbeat.interval.ms=60000
 debezium.source.topic.prefix=t
 debezium.source.tombstones.on.delete=false
-debezium.transforms=Reroute,ReadOperationFilter,PostsFilter
+debezium.transforms=Reroute,Notifications,ReadOperationFilter,PostsFilter
 debezium.transforms.Reroute.type=io.debezium.transforms.ByLogicalTableRouter
-debezium.transforms.Reroute.topic.regex=(.*)
+debezium.transforms.Reroute.topic.regex=^((?!\.notification_v2).)*$
 debezium.transforms.Reroute.topic.replacement=%topic%
+debezium.transforms.Notifications.type=io.debezium.transforms.ByLogicalTableRouter
+debezium.transforms.Notifications.topic.regex=.*\.notification_v2
+debezium.transforms.Notifications.topic.replacement=api.v1.cdc-notifications
 debezium.transforms.ReadOperationFilter.type=io.debezium.transforms.Filter
 debezium.transforms.ReadOperationFilter.language=jsr223.groovy
 debezium.transforms.ReadOperationFilter.condition=!(valueSchema.field('op') && value.op == 'r')
