+++ src/workers/index.ts
@@ -14,7 +14,8 @@ import postDeletedCommentsCleanup from './postDeletedCommentsCleanup';
 import usernameChanged from './usernameChanged';
 import sourceRequestApprovedRep from './sourceRequestApprovedRep';
 import updateComments from './updateComments';
-import cdc from './cdc';
+import cdc from './cdc/primary';
+import cdcNotifications from './cdc/notifications';
 import updateMailingList from './updateMailingList';
 import deleteUserFromMailingList from './deleteUserFromMailingList';
 import newNotificationRealTime from './newNotificationV2RealTime';
@@ -84,6 +85,7 @@ export const workers: Worker[] = [
   sourceSquadCreatedUserAction,
   sourceSquadCreatedOwnerMailing,
   cdc,
+  cdcNotifications,
   ...notificationWorkers,
 ];
